Add Cmd interface and run command
The Cmd interface allows for subcommands to be implemented as its own,
distinct structs.

The Run subcommand copies files either as specified by the user or by
reading a manifest. The manifest is a source->dest mapping. The copying
is done by calling the configured copying utility. There are many great
copying utility out there. There is no reason to reimplement the wheel
when we can just put tires on them instead.