Add generate subcommand
The Generate subcommand generates a manifest file from the user
specified src and dest. If the src is a directory, it is scanned and the
manifest will instead map its children to the dest. This is NOT done
recursively, so it is up to the copying utility to handle directories.