Add cleanup script
Cleanup is a complicated process as various scripts generates various
files for to prepare the environment for building release binaries.
While no one should ever build the _actual_ release binaries in their
local machine, it is an easy sanity check to run when the GH Actions
fails. Simplify the cleanup process.